Break SmfType into its own file.
